segments

+++ user.go
@@ -22,7 +22,7 @@ type User struct {
 	LastRequestAt          int32             `json:"last_request_at,omitempty"`
 	CreatedAt              int32             `json:"created_at,omitempty"`
 	UpdatedAt              int32             `json:"updated_at,omitempty"`
-	SessionCount           int64             `json:"session_count,omitempty"`
+	SessionCount           int32             `json:"session_count,omitempty"`
 	LastSeenIP             string            `json:"last_seen_ip,omitempty"`
 	SocialProfiles         SocialProfileList `json:"social_profiles,omitempty"`
 	UnsubscribedFromEmails *bool             `json:"unsubscribed_from_emails,omitempty"`
